+++ Projects/Domain/DataInterface/Sources/DeparturePlaceRepositoryProtocol.swift
@@ -8,7 +8,7 @@ import Entity
 public protocol DeparturePlaceRepositoryProtocol: Sendable {
     func fetchDeparturePlaces() async throws -> [DeparturePlace]
     func updateDeparturePlace(
-        id: Int64,
+        id: Int,
         label: String,
         address: String,
         roadAddress: String,
@@ -25,6 +25,6 @@ public protocol DeparturePlaceRepositoryProtocol: Sendable {
         longitude: Double,
         isDefault: Bool
     ) async throws -> DeparturePlace
-    func deleteDeparturePlace(id: Int64) async throws
-    func setDefaultDeparturePlace(id: Int64) async throws -> DeparturePlace
+    func deleteDeparturePlace(id: Int) async throws
+    func setDefaultDeparturePlace(id: Int) async throws -> DeparturePlace
 }
